requirements.txt
aiohttp>=3.8.0
fastapi>=0.103.0
python-dotenv>=1.0.0
streamlit>=1.26.0
uvicorn>=0.23.0
pydantic>=2.3.0
tavily-python>=0.2.0
python-uuid>=1.30
requests>=2.31.0
logging>=0.5.1.2
